+++ ETABS_Engine/Convert/ToBHoM.cs
@@ -33,15 +33,17 @@ using ETABS2016;
 using BH.oM;
 using BH.oM.Structure;
 using BH.oM.Structure.Elements;
+using BH.oM.Structure.Loads;
+using BH.oM.Structure.MaterialFragments;
+using BH.oM.Structure.Constraints;
 
 
 namespace BH.Engine.ETABS
 {
     public static partial class Convert
     {
-        /// <summary>
-        /// ids = null returns all nodes in model
-        /// </summary>
+        /***************************************************/
+
         public static List<Node> ToBHoM(this cPointObj pointObj, List<string> ids)
         {
             List<Node> bhNodes = new List<Node>();
@@ -62,6 +64,7 @@ namespace BH.Engine.ETABS
             return bhNodes;
         }
 
+        /***************************************************/
 
         public static Node ToBHoM(this cPointObj pointObj, string id)
         {
@@ -82,40 +85,98 @@ namespace BH.Engine.ETABS
             return bhNode;
         }
 
+        /***************************************************/
 
-        //public static Bar ToBHoM(this cFrameObj barObj, string id, ModelData modelData)
-        //{
-        //    Bar bhBar = new Bar();
-        //    bhBar.CustomData.Add(AdapterId, id);
-        //    string startId="";
-        //    string endId="";
-        //    barObj.GetPoints(id, ref startId, ref endId);
+        public static ICase CaseNameToBHoM(String csiCaseName, ICase bhomCase)
+        {
+            string[] nameNum = csiCaseName.Split(new[] { ":::" }, StringSplitOptions.None);
+            Int32 number = new Int32();
+            int.TryParse(nameNum[1], out number);
+            bhomCase.Number = number;
+            bhomCase.Name = nameNum[0];
 
-        //    //this method can only be called as 'model.frameObj.ToBHoM' 
-        //    //still 'model' needs to be passed as argument as well in order to get the nodes at bar ends
-        //    //this seems a bit convoluted way to keep to the convention of .ToBHoM extension methods !!!
+            return bhomCase;
+        }
 
-        //    bhBar.StartNode = modelData.model.PointObj.ToBHoM(startId);
-        //    bhBar.EndNode = modelData.model.PointObj.ToBHoM(endId);
+        /***************************************************/
 
-        //    bool[] restraintStart = new bool[6];
-        //    double[] springStart = new double[6];
-        //    bool[] restraintEnd = new bool[6];
-        //    double[] springEnd = new double[6];
+        public static LoadNature ToBHoM(this eLoadPatternType loadPatternType)
+        {
+            switch (loadPatternType)
+            {
+                case eLoadPatternType.Dead:
+                    return LoadNature.Dead;
+                case eLoadPatternType.SuperDead:
+                    return LoadNature.SuperDead;
+                case eLoadPatternType.Live:
+                    return LoadNature.Live;
+                case eLoadPatternType.Temperature:
+                    return LoadNature.Temperature;
+                case eLoadPatternType.Braking:
+                    return LoadNature.Accidental;
+                case eLoadPatternType.Prestress:
+                    return LoadNature.Prestress;
+                case eLoadPatternType.Wind:
+                    return LoadNature.Wind;
+                case eLoadPatternType.Quake:
+                    return LoadNature.Seismic;
+                case eLoadPatternType.Snow:
+                    return LoadNature.Snow;
+                default:
+                    return LoadNature.Other;
 
-        //    barObj.GetReleases(id, ref restraintStart, ref restraintEnd, ref springStart, ref springEnd);
-        //    bhBar.Release = new BarRelease();
-        //    bhBar.Release.StartRelease = GetConstraint6DOF(restraintStart, springStart);
-        //    bhBar.Release.EndRelease = GetConstraint6DOF(restraintEnd, springEnd);
+            }
+        }       
+        
+        /***************************************************/
 
-        //    eFramePropType propertyType = eFramePropType.General;
-        //    string propertyName = "";
-        //    string sAuto = "";
-        //    barObj.GetSection(id, ref propertyName, ref sAuto);
-        //    modelData.model.PropFrame.GetTypeOAPI(propertyName, ref propertyType);
-        //    bhBar.SectionProperty = GetSectionProperty(modelData, propertyName, propertyType);
-        //    return bhBar;
-        //}
+        private static MaterialType ToBHoM(this eMatType materialType)
+        {
+            switch (materialType)
+            {
+                case eMatType.Steel:
+                    return MaterialType.Steel;
+                case eMatType.Concrete:
+                    return MaterialType.Concrete;
+                case eMatType.NoDesign://No material of this type in BHoM !!!
+                    return MaterialType.Steel;
+                case eMatType.Aluminum:
+                    return MaterialType.Aluminium;
+                case eMatType.ColdFormed:
+                    return MaterialType.Steel;
+                case eMatType.Rebar:
+                    return MaterialType.Rebar;
+                case eMatType.Tendon:
+                    return MaterialType.Tendon;
+                case eMatType.Masonry://No material of this type in BHoM !!!
+                    return MaterialType.Concrete;
+                default:
+                    return MaterialType.Steel;
+            }
+        }
 
+        /***************************************************/
+
+        public static Constraint6DOF GetConstraint6DOF(bool[] restraint, double[] springs)
+        {
+            Constraint6DOF bhConstraint = new Constraint6DOF();
+            bhConstraint.TranslationX = restraint[0] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.TranslationY = restraint[1] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.TranslationZ = restraint[2] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.RotationX = restraint[3] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.RotationY = restraint[4] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.RotationZ = restraint[5] == true ? DOFType.Fixed : DOFType.Free;
+
+            bhConstraint.TranslationalStiffnessX = springs[0];
+            bhConstraint.TranslationalStiffnessY = springs[1];
+            bhConstraint.TranslationalStiffnessZ = springs[2];
+            bhConstraint.RotationalStiffnessX = springs[3];
+            bhConstraint.RotationalStiffnessY = springs[4];
+            bhConstraint.RotationalStiffnessZ = springs[5];
+
+            return bhConstraint;
+        }
+
+        /***************************************************/
     }
 }
