+++ Etabs_Adapter/Read/Node.cs
@@ -1,0 +1,92 @@
+/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2018, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using BH.oM.Base;
+using BH.oM.Structure.Elements;
+using BH.oM.Structure.SectionProperties;
+using BH.oM.Structure.SurfaceProperties;
+using BH.oM.Structure.Constraints;
+using BH.oM.Structure.Loads;
+using BH.oM.Structure.MaterialFragments;
+#if (Debug2017)
+using ETABSv17;
+#else
+using ETABS2016;
+#endif
+using BH.Engine.ETABS;
+using BH.oM.Geometry;
+using BH.Engine.Geometry;
+using BH.Engine.Reflection;
+using BH.oM.Architecture.Elements;
+using BH.oM.Adapters.ETABS.Elements;
+
+namespace BH.Adapter.ETABS
+{
+    public partial class ETABSAdapter
+    {
+        /***************************************************/
+
+        private List<Node> ReadNodes(List<string> ids = null)
+        {
+            List<Node> nodeList = new List<Node>();
+
+            int nameCount = 0;
+            string[] nameArr = { };
+
+            if (ids == null)
+            {
+                m_model.PointObj.GetNameList(ref nameCount, ref nameArr);
+                ids = nameArr.ToList();
+            }
+
+            foreach (string id in ids)
+            {
+
+                double x, y, z;
+                x = y = z = 0;
+                bool[] restraint = new bool[6];
+                double[] spring = new double[6];
+
+                m_model.PointObj.GetCoordCartesian(id, ref x, ref y, ref z);
+
+                m_model.PointObj.GetRestraint(id, ref restraint);
+                m_model.PointObj.SetSpring(id, ref spring);
+
+                Node bhNode = Engine.Structure.Create.Node(new oM.Geometry.Point() { X = x, Y = y, Z = z }, "", Engine.ETABS.Convert.GetConstraint6DOF(restraint, spring));
+                bhNode.CustomData.Add(AdapterId, id);
+
+                nodeList.Add(bhNode);
+            }
+
+
+            return nodeList;
+        }
+
+        /***************************************************/
+    }
+}
