+++ Etabs_Adapter/Types/NextIndex.cs
